Silence GCC 8 compiler warning in ShowEA()
The current version of GCC has gained the possibility to do some additional
checks for string operation buffer overflows. It is using the size of
source arrays for the check, so in ShowEA it is currently reporting (output
taken from Hatari build):

newcpu.c: In function 'ShowEA.constprop':
newcpu.c:2440:26: warning: '%s' directive writing up to 79 bytes into a
 region of size between 73 and 75 [-Wformat-overflow=]
    _stprintf (buffer, _T("(A%d, %s) == $%08x"), reg, offtxt, addr);

The offtxt array is only storing a string for a short displacement, so
80 characters is way too big here. Decreasing the size of the array to
less characters silences the compiler warning for me.