Silence compiler warning in get_long_cache_040
GCC with -O1 and Clang complain that 'v' may be used uninitialized in the
function get_long_cache_040(). The warning is wrong since all cases are
handled, so that v can never be used uninitialized here. But the last
check "if (offset == 14)" is redundant here anyway, so let's simply turn
this into an unconditional else-statement instead - which then also makes
the compilers happy.