Fix audio state machine stale intreq2 logic
As discussed in the "Undocumented Amiga hardware stuff" thread https://eab.abime.net/showpost.php?p=1679952&postcount=337
The audio state machine intreq2 signal should only be cleared when used to request an interrupt (AUDxIR).

Previously, intreq2 was cleared (and no interrupt requested) on the transition from state 5 to state 1.
Now, intreq2 is checked on transition from state 0 to state 1. If it is set, then an interrupt is requested and intreq2 is cleared.

Logically this is the same for the internal state machine, but the interrupts have been shown to be present on silicon using test case https://github.com/dirkwhoffmann/vAmigaTS/blob/master/Paula/Audio/timing/dmatim2/ and with this change, three missing interrupts are now generated in this test case (I'll add some images to the associated pull request).